Add manual/ci run mode for e2e tests run command
When switching between localhost and github.com for testing I sometimes get issues with the known_hosts file. Thus I make it more visible. If you pass the --manual flag (this flag is default) it tests the ssh connection. This will show if there are any errors. On the CI we don't want that, that is why we skip it on the runner.